Tests execution ordering
